+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [3.0.1] - Unreleased
 ### Added
 - Restore `site.copy()` function removed in 3.0.0.
-- Info from minify_html, esbuild, terser, postcss and svgo plugins to the debugbar.
+- Info from minify_html, esbuild, terser, postcss, lightningcss, google_fonts and svgo plugins to the debugbar.
 
 ### Fixed
 - Warning message by the transform_images plugin.